memory-agent/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "memory-agent"
version = "0.1.0"
description = "Privacy-first local cognitive memory agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Memory Agent Contributors"}
]
keywords = ["memory", "ai", "local", "privacy", "rag", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pylance>=0.10.0",
    "sentence-transformers>=2.3.0",
    "llama-cpp-python>=0.2.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.1.0",
]

all = [
    "faster-whisper>=0.9.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
]

[project.scripts]
memory = "memory_agent.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/memory-agent"
Documentation = "https://github.com/yourusername/memory-agent/docs"
Repository = "https://github.com/yourusername/memory-agent"
Issues = "https://github.com/yourusername/memory-agent/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=memory_agent --cov-report=term-missing"
